+++ lib/dealog_backoffice_web/live/organization_messages/edit.ex
@@ -83,11 +83,53 @@ defmodule DealogBackofficeWeb.OrganizationMessagesLive.Edit do
         )
         |> push_redirect(to: Routes.organization_messages_path(socket, :show, id))
 
+      {:error, :not_found} ->
+        socket
+        |> put_flash(
+          :save_error,
+          gettext("Message %{id} could not be found",
+            id: id
+          )
+        )
+        |> push_redirect(to: Routes.organization_messages_path(socket, :show, id))
+
       {:ok, message} ->
         socket
         |> put_flash(
           :save_success,
           gettext("Message %{title} successfully deleted", title: message.title)
+        )
+        |> push_redirect(to: Routes.organization_messages_path(socket, :index))
+    end
+  end
+
+  defp apply_action(socket, :archive, %{"id" => id}) do
+    case Messages.archive_message(id) do
+      {:error, :invalid_transition} ->
+        socket
+        |> put_flash(
+          :save_error,
+          gettext("Message %{id} could not be archived",
+            id: id
+          )
+        )
+        |> push_redirect(to: Routes.organization_messages_path(socket, :show, id))
+
+      {:error, :not_found} ->
+        socket
+        |> put_flash(
+          :save_error,
+          gettext("Message %{id} could not be found",
+            id: id
+          )
+        )
+        |> push_redirect(to: Routes.organization_messages_path(socket, :show, id))
+
+      {:ok, message} ->
+        socket
+        |> put_flash(
+          :save_success,
+          gettext("Message %{title} successfully archived", title: message.title)
         )
         |> push_redirect(to: Routes.organization_messages_path(socket, :index))
     end